refactor: useTimerPageState에서 모달 관련 훅을 분리
